pdf_to_md_app/requirements.txt
Flask==3.0.3
unstructured[pdf]>=0.16.19
unstructured-inference>=0.8.0
openai==1.14.2
httpx<0.28.0
tiktoken>=0.9.0
python-dotenv==1.0.1
